Fix types to avoid warnings.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gccgo@157066 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgo/runtime/mheap.c
@@ -241,7 +241,7 @@ MHeap_FreeLocked(MHeap *h, MSpan *s)
 	MSpan *t;
 
 	if(s->state != MSpanInUse || s->ref != 0) {
-		printf("MHeap_FreeLocked - span %p ptr %zu state %d ref %d\n", s, s->start<<PageShift, s->state, s->ref);
+		printf("MHeap_FreeLocked - span %p ptr %zu state %u ref %u\n", s, (size_t) (s->start<<PageShift), (unsigned int) s->state, (unsigned int) s->ref);
 		throw("MHeap_FreeLocked - invalid free");
 	}
 	s->state = MSpanFree;
